refactor(signalk-server): retire the runtime provisioning hook
The plugins ship in the image now, so the hook, its manifest, its harness and
the one-shot unit it drove have nothing left to do. Removed rather than left
inert: a hook that still runs would reinstall all 17 packages into the data
volume, where they shadow the baked copies for good.

The manifest's tests go with it. Their job was to stop a typo from withholding
the app on every device -- a failure mode that no longer exists, because a bad
entry now fails the image build instead of a boat. The one that did not
transfer, test_influxdb_plugin_is_present, moves to the image repo alongside the
manifest it guards (halos-org/signalk-server-docker#2): prestart.sh writes that
plugin's config unconditionally, so dropping it from the curated set would stop
marine logging with nothing failing.

tests/test_signalk_plugins.py is renamed to match what is left in it, which is
prestart.sh and nothing else. .gitattributes went with the manifest it pinned.
tools/verify-payloads.sh keeps its generic provisioning branch; grafana's assets
keep the check non-vacuous, and the branch stays correct until the generator
drops the feature.
